Move connection and inserter into a shared db directory.
