+++ lsif/src/server/backend/cache.ts
@@ -1,7 +1,7 @@
 import promClient from 'prom-client'
 import Yallist from 'yallist'
 import { Connection, EntityManager } from 'typeorm'
-import { createSqliteConnection } from '../../connection'
+import { createSqliteConnection } from '../../shared/database/sqlite'
 import { DocumentData, ResultChunkData } from '../../database.models'
 import {
     connectionCacheEventsCounter,
