+++ lsif/src/server/server.ts
@@ -12,7 +12,7 @@ import { createJobRouter } from './routes/jobs'
 import { createLogger } from '../shared/logging'
 import { createLsifRouter } from './routes/lsif'
 import { createMetaRouter } from './routes/meta'
-import { createPostgresConnection } from '../connection'
+import { createPostgresConnection } from '../shared/database/postgres'
 import { createQueue, ensureOnlyRepeatableJob } from '../queue'
 import { createTracer } from '../shared/tracing'
 import { dbFilename, dbFilenameOld, ensureDirectory } from '../util'
