+++ lsif/src/shared/xrepo/xrepo.ts
@@ -9,7 +9,7 @@ import { instrument } from '../metrics'
 import { Connection, EntityManager, Brackets } from 'typeorm'
 import { createFilter, testFilter } from '../../encoding'
 import { PackageModel, ReferenceModel, Commit, LsifDump, DumpId } from '../../xrepo.models'
-import { TableInserter } from '../../inserter'
+import { TableInserter } from '../database/inserter'
 import { addrFor, getCommitsNear, gitserverExecLines } from './commits'
 import { TracingContext, logAndTraceCall } from '../tracing'
 import { dbFilename, tryDeleteFile } from '../../util'
