+++ lsif/src/test-utils.ts
@@ -2,7 +2,7 @@ import * as fs from 'mz/fs'
 import * as path from 'path'
 import * as uuid from 'uuid'
 import { Connection } from 'typeorm'
-import { connectPostgres } from './connection'
+import { connectPostgres } from './shared/database/postgres'
 import { lsp } from 'lsif-protocol'
 import { child_process } from 'mz'
 import { convertLsif } from './worker/importer/importer'
