+++ lsif/src/worker/worker.ts
@@ -6,7 +6,7 @@ import { createCleanFailedJobsProcessor } from './processors/clean-failed-jobs'
 import { createCleanOldJobsProcessor } from './processors/clean-old-jobs'
 import { createConvertJobProcessor } from './processors/convert'
 import { createLogger } from '../shared/logging'
-import { createPostgresConnection } from '../connection'
+import { createPostgresConnection } from '../shared/database/postgres'
 import { createQueue } from '../queue'
 import { createUpdateTipsJobProcessor } from './processors/update-tips'
 import { ensureDirectory } from '../util'
